+++ crates/async-components/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name         = "ibc-relayer-runtime"
+name         = "ibc-async-components"
 version      = "0.1.0"
 edition      = "2021"
 license      = "Apache-2.0"
@@ -18,13 +18,10 @@ all-features = true
 [features]
 
 [dependencies]
-async-trait = "0.1.56"
-futures = "0.3"
-tokio = "1.24"
-tokio-stream = "0.1"
-tracing = "0.1.36"
-flex-error = { version = "0.4.4", default-features = false }
-cgp-core = { version = "0.1.0" }
 ibc-relayer-components = { version = "0.1.0", path = "../relayer-components" }
 ibc-relayer-components-extra = { version = "0.1.0", path = "../relayer-components-extra" }
-ibc-async-components = { version = "0.1.0", path = "../async-components" }
+cgp-core = { version = "0.1.0" }
+cgp-macros = { version = "0.1.0" }
+async-trait = "0.1.56"
+futures-core = { version = "0.3", default-features = false, features = ["alloc"] }
+futures-util = { version = "0.3", default-features = false, features = ["alloc"] }
