+++ crates/async-components/src/impls/subscription/empty.rs
@@ -3,8 +3,9 @@ use core::pin::Pin;
 
 use async_trait::async_trait;
 use cgp_core::traits::Async;
-use futures::stream::Stream;
+use futures_core::stream::Stream;
 
+use crate::std_prelude::*;
 use crate::traits::subscription::Subscription;
 
 pub struct EmptySubscription<T>(pub PhantomData<T>);
