+++ crates/async-components/src/impls/subscription/multiplex.rs
@@ -5,13 +5,15 @@ use ibc_relayer_components::runtime::traits::stream::HasStreamType;
 
 use async_trait::async_trait;
 use cgp_core::traits::Async;
-use futures::stream::{Stream, StreamExt};
+use futures_core::stream::Stream;
+use futures_util::stream::StreamExt;
 use ibc_relayer_components::runtime::traits::mutex::HasMutex;
 use ibc_relayer_components_extra::runtime::traits::channel::{
     CanCreateChannels, CanStreamReceiver, CanUseChannels, HasChannelTypes,
 };
 use ibc_relayer_components_extra::runtime::traits::spawn::{HasSpawner, Spawner};
 
+use crate::std_prelude::*;
 use crate::traits::stream::HasAsyncStreamType;
 use crate::traits::subscription::Subscription;
