+++ crates/async-components/src/lib.rs
@@ -1,6 +1,8 @@
+#![no_std]
 #![allow(clippy::type_complexity)]
 
+mod std_prelude;
 extern crate alloc;
 
 pub mod impls;
-pub mod types;
+pub mod traits;
