+++ crates/relayer-cosmos/Cargo.toml
@@ -28,6 +28,7 @@ ibc-relayer-runtime = { version = "0.1.0", path = "../relayer-runtime" }
 ibc-relayer-all-in-one = { version = "0.1.0", path = "../relayer-all-in-one" }
 ibc-relayer-components = { version = "0.1.0", path = "../relayer-components" }
 ibc-relayer-components-extra = { version = "0.1.0", path = "../relayer-components-extra" }
+ibc-async-components = { version = "0.1.0", path = "../async-components" }
 
 serde = "1.0"
 serde_derive = "1.0"
