+++ crates/relayer-cosmos/src/impls/chain.rs
@@ -1,6 +1,7 @@
 use alloc::sync::Arc;
 
 use async_trait::async_trait;
+use ibc_async_components::traits::subscription::Subscription;
 use ibc_relayer::chain::client::ClientSettings;
 use ibc_relayer::chain::endpoint::ChainStatus;
 use ibc_relayer::chain::handle::ChainHandle;
@@ -10,7 +11,6 @@ use ibc_relayer::event::{
 };
 use ibc_relayer_all_in_one::one_for_all::traits::chain::{OfaChain, OfaChainTypes, OfaIbcChain};
 use ibc_relayer_components::chain::traits::components::message_sender::CanSendMessages;
-use ibc_relayer_runtime::traits::subscription::Subscription;
 use ibc_relayer_runtime::types::error::Error as TokioError;
 use ibc_relayer_runtime::types::log::logger::TracingLogger;
 use ibc_relayer_runtime::types::log::value::LogValue;
