+++ crates/relayer-runtime/src/impls/mod.rs
@@ -1,4 +1,3 @@
 pub mod logger;
 pub mod runtime;
-pub mod subscription;
 pub mod task;
