+++ crates/relayer-runtime/src/impls/runtime/stream.rs
@@ -3,9 +3,9 @@ use core::pin::Pin;
 use cgp_core::traits::Async;
 use futures::stream::Stream;
 use futures::StreamExt;
+use ibc_async_components::traits::stream::HasAsyncStreamType;
 use ibc_relayer_components::runtime::traits::stream::{CanMapStream, HasStreamType};
 
-use crate::traits::stream::HasAsyncStreamType;
 use crate::types::runtime::TokioRuntimeContext;
 
 impl HasStreamType for TokioRuntimeContext {
